Fix CodeQL warnings (#38)
* Fix codeQL warnings

* update codeowners

+++ test/BenchmarkDotNetAnalyser.Tests.Unit/Benchmarks/BenchmarkStatisticAccessorProviderTests.cs
@@ -151,7 +151,7 @@ namespace BenchmarkDotNetAnalyser.Tests.Unit.Benchmarks
 
             var result = a(br);
 
-            result.Should().Be((decimal)expected);
+            result.Should().Be(expected);
         }
 
         [Theory]
